chore(os): Drop xenial support for eol
Drop support for Ubuntu Xenial as it reached its EOL.
This means upgrading Travis jobs and Vagrant box too.

Also, there is an interesting observation that git does not print tag
information in the format introduced in #1768 if the tag is on current
commit. This causes Debian builds to fail.
The solution is to make format more flexible and add additional 0
revision.

Signed-off-by: Gaurav Mishra <mishra.gaurav@siemens.com>